--- src/mono/mono/utils/networking-windows.c
@@ -1,34 +1,0 @@
-/**
- * \file
- * Windows-specific networking implementations
- *
- * Author:
- *	Alexander Köplinger (alex.koeplinger@outlook.com)
- */
-
-#include <mono/utils/networking.h>
-
-#if defined(HOST_WIN32)
-
-void
-mono_networking_init (void)
-{
-	WSADATA wsadata;
-	int err;
-
-	err = WSAStartup (2 /* 2.0 */, &wsadata);
-	if(err)
-		g_error ("%s: Couldn't initialise networking", __func__);
-}
-
-void
-mono_networking_shutdown (void)
-{
-	WSACleanup ();
-}
-
-#else
-
-MONO_EMPTY_SOURCE_FILE (networking_windows);
-
-#endif /* defined(HOST_WIN32) */
